move choropleth legend slider into choropleth legend directory.
