+++ src/ui/choropleth-legend/src/slider-handle.css
