+++ src/ui/choropleth-legend/src/slider-handle.jsx
