+++ src/ui/choropleth-legend/src/slider.jsx
