+++ src/ui/choropleth-legend/test/slider-handle.test.js
