+++ src/ui/choropleth-legend/test/slider.test.js
