--- src/ui/slider/index.js
@@ -1,1 +1,0 @@
-export { default as default } from './src/slider';
