--- src/ui/slider/test/.eslintrc.js
@@ -1,5 +1,0 @@
-module.exports = {
-  "rules": {
-    "no-unused-expressions": 0
-  }
-};
